add Status entiti and set relationship 1-1 with dream
